requirements.txt
discord.py ~= 1.7.3
pandas ~= 1.2.4
odfpy ~= 1.4.1